Re-work 1.4-packages to add tier and state fields (#1123)
By updating the package list specification to specify the target tier as
well as the desired state (present or absent) this allows us to specify
both packages we want installed, as well as packages we want removed,
for a given target tier. Note that an 'all' tier entry will always be
considered a match for the active target tier.

When calling the package module we extract the matching list of packages
for a given state into a list to be passed as the name parameter per the
recommended calling pattern; this minimises the calls to the underlying
package manager.

Add definitions for 'distro_name' and 'distro_id' to the 1.4-packages
role defaults to make it easier to read the code.
NOTE: This is an example of the sort of definitions that it is useful
to include in a group_vars/all file, making then available to every
playbook, role, and task without having to specify them or explicitly
include them.

Add target tier specification to the top level playbook associated with
the roles-os roles.

Add support to test_menu.sh to pass any arguments specified when running
it as additional arguments when calling ansible-playbook. For example
this allows verbosity control for the running ansible playbook.