Implement 1.11-accounts (#1130)
Implement a role that can be used to create admin account, and
associated groups, with matching /etc/sudoers.d/<username> files.

The role supports both adding and removing admin users and groups,
as well as allowing account settings to be associated with a given
playbook tier, and enabled/disabled via the associated attribute.

The details of the admin groups and accounts to be created are specified
in the top level vars/admin-accounts.yml file. Currently this file holds
plain text passwords and static uid and gid definitions; this needs to
change to creating/retrieving passwords from a secure source such as a
keyvault, and allow for dynamic uid & gid assignments which are tracked
in a persistent fashion in a keyvault.

Currently just two "flavours" of sudoers roles are defined:
  * admin_no_password:
    all commands permitted without need to enter a password.
  * admin_password_required:
    all commands permitted but password validation required.

Additional flavours can be added by adding new templates with a name
in the format 'sudoers_<role_name>.j2'. Optionally a vars file with
the same name can be created, which will be dynamically loaded if
found, that can hold any useful fact definitions that may aid in
processing the matching template file.