Fix playbook_01 tags enablement (#1163)
Previous change just added tags to the include_role actions, but this
approach only works for import_role. For include_role the tags apply
to the include_role statement itself, and not to the included tasks
so unless they also have the same tags configured, or are tagged with
always, they won't be run.

To get the desired effect, namely run only the tasks associated with
the included role when the associated task is specified, we need to
specify both a tags list for the include_role, and dynamically apply
the same tags to the included tasks using the apply mechanism. Or
you can wrap the include_role in a block action, with associated
tags, which will apply to both the include_role action and any tasks
that it includes.

Co-authored-by: Kimmo Forss <kimmo.forss@microsoft.com>