Add --only_deployer option to prepare_region.sh (#1239)
When the --only_deployer option is used with the prepare_region.sh
script it causes to stop once the deployer has been successfully
created. At this point the user can now copy the relevant config
and state files to the deployer and run prepare_region.sh from there
to complete to region setup.